[scripts/ci/select_tezts.sh]: make merge train safe
Diff [CI_MERGE_REQUEST_DIFF_BASE_SHA] with
[CI_MERGE_REQUEST_SOURCE_BRANCH_SHA] in merge train pipelines.

Merge train pipelines run on commits that are:

 - rebased on the target branch
 - may contain the commits of other MRs in the merge traine.

Furthermore, the variable [CI_MERGE_REQUEST_DIFF_BASE_SHA] is not
updated in such pipelines to point to its intuitive value (the most
recent commit on the current branch which does not come from the MR
itself). Instead, it contains the same value as for the un-rebased MR.

Consequently, the comparison between [HEAD] and
[CI_MERGE_REQUEST_DIFF_BASE_SHA] will be larger than intended,
containing the diff of the MRs merged since the MRs rebase, and the
diff of preceding MRs in the merge train.

This is solved by diffing [CI_MERGE_REQUEST_DIFF_BASE_SHA] with
[CI_MERGE_REQUEST_SOURCE_BRANCH_SHA] in merge train pipelines. The
later variable points to the HEAD of the MRs source branch.