Etherlink/Sequencer/Tests: increase timeouts for DAL latency
This is a preparation for DAL integration in Etherlink. We need to
increase the timeouts in order to wait enough blocks for the
blueprints to be sent through the DAL. The added latency is:

- possibly one block for the rollup node to inject on the DAL,
- 8 blocks of lag to let DAL attesters attest the DAL slot (this is a
  DAL protocol parameter called "attestation_lag"),
- 1 extra block before the slot can be imported

Moreover, in !13883, the sequencer sends at most one blueprint per L1
level on the DAL so, to catch up, a proxy EVM node which is N L2
levels behind the sequencer needs N extra L1 levels. Since the tests
relevant to the DAL usually have a sequencer 5 levels ahead, we
increase the timeouts by 15 levels (from 5 to 20).