DAL/Tests: simplify test
... by publishing only at the required level

Also wait for the producer and attester to be part of the mesh at the right
level, and also wait for the attester to receive its shards. This should
eliminate one source of flakiness, where if the graft does not happen at the
right level or the shards are not received in time, the slot cannot be attested.