EVM Node: Expose the maximum time between two blocks via a RPC
We introduce `GET /evm/time_between_blocks` in order for observer nodes to
(recursively) get the maximum time between two blocks from the sequencer. As a
consequence, we can deprecate the `time_between_blocks` field of the observer
configuration. As a consequence, the detection of stalled streamed RPCs becomes
the default behavior for observer nodes, while it was not the case before.

In order to avoid an unnecessary breaking changes from operators perspective,
we keep the field `time_between_block` as part of the encoding definition of
the configuration type, but it is being fully ignore.