EVM Node: Introduce the `kernel_execution` configuration field
Instead of duplicated `preimages` and `preimages_endpoint` in modes that
require it, we rather put everything in `kernel_execution`. This is possible
because it is quite easy to have a default value solely based on `data_dir`. As
a consequence, the `kernel_execution` field can be omitted for the proxy node
without any issue.