EVM Node: The RPC mode does not rely on the observer config anymore
The only remaining usage of the observer config by the RPC mode is to get a EVN
node endpoint to connect to. Instead, we compute an EVM endpoint based on the
`rpc_port` and `rpc_addr` config value.

The current proposal will assume http:// as the scheme to connect to the
read-write node, which is consistent with every deployment we have from
partners. Besides, it is compatible with what I’m calling “catch-all bind
addresses” in both IPv4 and IPv6. When these are used as bind address, we
replace `rpc_addr` with localhost as the way to connect to the read-write node.