EVM Node: Proper blueprints streaming from RPC nodes
In order to do that, we need for the RPC mode to follow the head of the running
read-write node. In order to do that, we extract from the `Observer` module the
logic of following the head of a remote EVM node. We use this now generic logic
to keep up to date a “minimal” head information type, only consisting in the
next blueprint number.

Besides, we make the `Evm_services` module independant from `Evm_context` with
dependency injection: that is, the function that are used to interact with the
context are now passed as argument of `Evm_context.register`. This way, the RPC
mode can rely on ad-hoc function which do not require starting the
`Evm_context` worker.