Etherlink: move safe gas estimation to very first call
These are safe amounts to add to the gas. They are needed
by the execution but are missing from outcome gas used.
The code used to add the safety amounts everytime the kernel was
called, this is incorrect.

We now do it once. If it's enough, the gas is confirmed. If not,
the gas is doubled.