RPC: FIX: RPC middleware and resto use call_with_closefn() and sleep_fn()
Co-authored-by: Diana Savvatina <diana.savvatina@trili.tech>

(Diana Savvatina was the main author of this, I merely rebased)

Adding a callback for sleep_fn() requesting Cohttp to wait for EOF
from the client.

Adding a callback for conn_closed() allowing
Cohttp to notify the RPC middlewhere if the client
closes the connection or dies.

RPC middleware requests and saves the close_fn callback which is
called when the created connection should be closed.