EVM Node: Disable kernels safety check of the Fast Execution runtime
The Fast Execution runtime of the WASM PVM is keeping a cache of parsed module,
to avoid the penalty of parsing WASM programs over and over. From the
perspective of the EVM node, it means every time a RPC is called targeting a
block using a kernel not already seen by the node, this kernel will be parsed.

In 40721cae3572f5aabbba1de4a37f441528fdeca2, we have made things worst by
making the Fast Execution verify that the kernel to be compiled satisfy the
specific constraint of the WASM PVM (mostly, that it does not use floats at
all).

This safety check is absolutely necessary for the rollup node, because a faulty
kernel can lead a honest operator to loose its stake. But for the EVM node,
this safety measure makes less sense. As a consequence, we disable it. On my
test machine, the penalty when loading a brand new kernel for the first time
goes down from around 7s to 2s.