EVM Node: Preload known kernels at start-up
When the Fast Execution runtime sees a kernel for the first time, it needs to
parse it. This can be costly, especially in the context of an indexing process
making many requests at the same time (in that case, the kernel will be parsed
several times).

Fortunately, we have everything we need in the SQLite database to know where to
enumerate all the kernels from the past that are susceptible to be parsed, by
relying on the `Kernel_upgrades` table. This table list the activation level of
the previous upgrades: we can use this to know which EVM state can be used to
preload our cache.