Docker: Install Opam 2.2 from Alpine Edge repository
Listen, I haven’t chosen this. It is not my fault if our dependency tree now
requires us to have Opam 2.2 installed. Blame the most recent version of
`ocaml-env-msvc32.1` (I think). Me? I just want to enjoy the latest version of
`ocaml-lsp-server`. Which is not part of the opam repository hash we are
relying on today.

Anyway, fortunately, Opam 2.2 is already packages for Alpine, in the so-called
Edge repository. I guess this means Opam 2.2 will be part of the next Alpine
release, but considering Alpine 3.20 is a few weeks old, I propose to take a
shortcut in the meantime and install Opam from the Edge repository.

You can argue `ocaml-lsp-server` is a shallow reason to do such a thing, but
the next time we need to upgrade an external dependency in a hurry, you might
be happy to not have to deal with Opam 2.2. Not to say the stuggle is
completely over. We are still inhereting from a conflict in our dependency
tree. But at least, Opam is no longer an issue.