EVM Node: Remove useless patching of the config (RPC mode)
The way the RPC mode deals with patching on-disk configuration with the CLI is
weird at the moment. This is due to a comment left during a review that was
taken into account, but with some undesired leftovers.

It’s not like if the previous code was incorrect, but the manual patching that
is removed by this patch is fully redundant with the call to
`patch_configuration_from_args` happening earlier in the fuction body.