EVM Node: Fix eth_callTrace not going through `Backend_rpc`
The functor hierarchy of the EVM node is rather annoying to explain, but in a
nutshell, the RPCs are not supposed to use the `Tracer` module directly, but
rather go through the `Backend_rpc` first-class module.

With the current implementation, a call to `debug_traceCall` to a Proxy node
will raise an error about the `Evm_context` not being started instead of a
mention that this RPC is not supported.