CI: temporary allow 'commit_title' to fail (for any exit code)
This job checks, running the script
[scripts/ci/check_commit_messages.sh], that the commit titles do not
contain certain patterns described in
[scripts/ci/forbidden_commit_messages.txt].

This script will return exit code 65 if any of the MR commit titles
contain forbidden patterns.

Currently, the job is allowed to fail for this exit code only; which means
that this job should either succeed (titles are ok) or pass with a
warning (titles are ko).

Due to a bug (cf. issue https://gitlab.com/tezos/tezos/-/issues/7436),
this job will fail with exit code 1 for merge request pipelines.

This MR allows this job to fail whatever the reason. It is a temporary
fix before finalising the patch for [scripts/ci/check_commit_messages.sh].