EVM Node: Allow to limit the maximum size of a JSON RPC API batch
The JSON RPC API is a funny beast, which allow in a single HTTP Request to
perform any call to the API. This can be a problem when these calls are
computationally significant. In order to mitigate risks when this happens, we
introduce a new configuration parameter that allows to set a bound on the size
of one batch.

The current behavior is as follows: if a batch is larger than the authorized
limit, then all the calls fail. To be compliant with the [JSON RPC API
specification], we return one failure per call, and they all have the same
error. We might want to revisit this to allow at most `N` calls, where `N` is
the limit authorized.

[JSON RPC API specification]: https://www.jsonrpc.org/specification#batch