EVM Node: Enable the RPC mode to answer tracing RPCs
The current implementation of the tracing RPCs directly depends on the
`Evm_context` module. This module is responsible for the consistency between
the SQLite store and the Irmin repo of the EVM node, and acquires a read-write
“lock” on it. As a consequence, the RPC mode does not start the `Evm_context`
worker, and cannot answer these RPCs.

It turns out that you don’t need write permission to replay blocks and
transactions, and the current implementation was made this way mostly out of
convenience. This patch proposes a minimal (but not necessarily very
satisfactory) fix to enable the RPC mode to answer these RPCs.