EVM Node: Support observer nodes not tracking a rollup node
Since we stream kernel upgrades through blueprints, it is actually possible now
to have an observer node only connected to another EVM node endpoint. We turn
this theoritical possibility into a reality by adding a
`--dont-track-rollup-node` switch argument to the observer mode (along with a
companion `track-rollup-node` field in the observer configuartion).

The trust model is obviously different, but this can make it significantly
easier to setup a new node. Assuming the EVM node endpoint the observer is
connected to a rollup node, the trust model does not fully rely on the
sequencer operator.