Tezt: Use a RPC node in `Evm_rollup` scenarios instead of the sequencer
This is the easiest way to we ensure the EVM node in RPC mode can answer to
RPCs correctly. Besides, considering we want to migrate to a world where the
RPC mode is the unique mode that really answers RPC calls directly, this is a
step in this direction as well.