EVM Node: Keep the result of `eth_blockNumber` in memory
Compared to the sequencer and observer modes, the RPC mode does not keep the
head of the chain in memory, and instead rely the store and the Irmin context
extensively. The main reason for this was that SQLite was reasonably fast, and
that Irmin keeps a cache that we don’t clear (because we don’t commit).

That being said, as part of the Infra V2 project, we need to enable the
`eth_blockNumber` RPC for the relay.network.etherlink.com enpdoint. We really
want this endpoint to be as light as possible, and if we can make
`eth_blockNumber` almost free, so be it.

It is a good thing we can!