EVM Node: Refactor how RPC servers are started
Let’s face it folks. The EVM node is a mess, in a few places. The `main`
functions of our various execution modes are a good place to start. Have you
ever read the initialization and finalization of our RPC servers and fled,
tears in your eyes?

Well, no more!

We introduce `Rpc_server.start`, a function which we the use in `proxy.ml`,
`sequencer.ml`, `threshold_encryption_sequnecer.ml`, `observer.ml` and `rpc.ml`
to start the public and private servers. We use this opportunity to enable the
private RPC server for observer nodes as well.

Isn’t it nice? I think it’ nice.