Tezt: Fix flakiness of “Alpha: RPC method eth_getBlockReceipts”
On my machine, the test is often flaky, and it is because the sequencer
sometimes creates a new block while the transactions are being injected.

Setting the maximum between two blocks to `none` allows to avoid this pitfall,
and works because the helpers used in this test do not assume automatic block
creation and call the private RPC when necessary.