EVM/Execution: [fix] wrong depth for CREATEs for call tracing
We need to wait for the end_inter_transaction in order to have a stack
depth that is coherent and consistent like for calls.
Otherwise the transaction layer data isn't pop and the depth is off by one.

NB: The diff outputed by git is kind of misleading.
I've just moved the code for the tracing of creates after the
end_inter_transaction function.