EVM/Execution: [fix] remove the revert reason field from tracing
The reason for that is simple, it's the output from the machine.
We already have that field so no need for this field here which was
filled with placeholders anyway.

This commit is the kernel part. The node needs to also decode the output
as a UTF8 string to get the proper value.