EVM Node: Pop transaction objects from the Tx Pool
The next step for the RPC mode is to inject transaction object directly to the
upstream EVM node it is connected to, effectively by-passing the transaction
validation routine for this node.

The thing is, as it is implemented now, the transaction objects do not leave
the Tx Pool when they are poped: only raw transactions are. This patch fixed
that, as a preliminary step. As a nice side effect, this means the observer and
RPC modes does not have to recompute a hash that is already available to them.