EVM Node: Fix `Evm_store.get_activation_levels` request
It is possible to have an entry in the kernel upgrades table which does not
have `applied_before` set. In such a case, the SQL statement will fail, because
it will lead Caqti to try to cast `NULL` to an integer.

This simple one-liner fix will remove such a row from the results.