EVM Node: Fix the uniqueness constraint for the `kernel_upgrades` table
SQLite treats `NULL` in a particular manner, in that `NULL != NULL`. In order
to have a real uniqueness constraint for `applied_before`, we need to replace
`NULL` by a dummy value. Considering `applied_before` is a level, then `-1` is
a valid dummy value.

I was able to confirm this constraint works by manipulating a SQLite database
where the EVM node migrations were applied.

    sqlite> select * from migrations order by id desc limit 1;
    9|applied_before_uniqueness
    sqlite>  REPLACE INTO kernel_upgrades (injected_before, root_hash, activation_timestamp) VALUES (1, 'test', 111);
    sqlite>  REPLACE INTO kernel_upgrades (injected_before, root_hash, activation_timestamp) VALUES (2, 'test2', 1011);
    sqlite> select * from kernel_upgrades ;
    2|test2|1011|

To be compared with the previous behavior.

    sqlite> select * from migrations order by id desc limit 1;
    8|reworking_l1_l2_levels_table
    sqlite>  REPLACE INTO kernel_upgrades (injected_before, root_hash, activation_timestamp) VALUES (1, 'test', 111);
    sqlite>  REPLACE INTO kernel_upgrades (injected_before, root_hash, activation_timestamp) VALUES (2, 'test2', 1011);
    sqlite> select * from kernel_upgrades ;
    1|test|111|
    2|test2|1011|