EVM Node: Print SQLite schemas
It can be difficult to keep track of the current SQLite tables, considering we
only keep successive migration scripts. Arguably, it is possible to infer the
current schemas by reading the SQLite statements in `evm_store.ml`.

We propose a more direct way, by introducing a new command `debug print store
schemas`. This one is not intented for end-users, but rather as a development
tool. We add a regression test which can be used to review the current schemas
without having to remember the command.