Deps: Bump ocamlformat.0.26.2 to be compatible with OCaml 5.2
We are using `ocamlformat.0.25.1` which has been released on March, 2023. OCaml
is now a fast moving language, which regular breaking changes. For instance, if
we want to use the latest version of `ocaml-lsp-server` (1.19.0), we need OCaml
5.2. But if we want OCaml 5.2, we need `ocamlformat` 0.26.2.

So here we are.

In general, we should strive to remain up to date with our language of choice
toolchain anyway.