EVM Node: RPC servers configuration big bang
The configuration of the two RPC servers exposed by the EVM node (public, and
private) is currently very archaic. We propose a refactoring that should
improve the situation, by defining a clean `rpc` type gathering all the
configuration options that are currently at top-level.

A pretty convenient by-product of this approach is to provide the same features
for the private RPC server than the public ones (including ACL for the JSON RPC
API service).