EVM Node: Remove deprecated config options and improve documentation
We have tried to keep the EVM node configuration file somewhat retro-compatible
for a while now. The result is a joyful mess, where old fields are ignored but
do not prevent decoding.

Besides, we have seen some inconsistencies between the use of - and _. This
makes the configuration file very hard to maintain over time.

With this patch, we propose three things.

  1. We clean-up the schema by removing deprecated fields, and using _
     consistently.
  2. We introduce a new command `describe config` which prints the JSON schema
     to the standard output. We use the opportunity to document most fields,
     and to provide stricter encoding when it makes sense (for instance, when
     integers are expected to be strictly positive, or when we know we want
     human readable strings).
  3. We introduce a regression test to help us keep the configuration schema
     readable and informative.