EVM Node: Ensure that we always uses the latest WASM PVM version
With the DAL+Etherlink project, the kernel now has a dependency to the third
revision of the WASM PVM. Turns out, the sequencer of the testnet is still
using the second one.

It should be fine to always use the latest version of the PVM, because in the
end, what matters for the safety of the chain is the version used by the rollup
nodes. Our test-suite is resilient enough to catch when we use a feature that
is not available in the protocol we care about (since we execute Etherlink
tests on the current protocol).