CI/Docker: prefer publishing Octez Docker distribution to Docker Hub
Switches order of the blocks for authenticating with the GCP registry
resp Docker Hub. Such that, if credentials for Docker Hub are
available (as on protected refs), then [docker_image_name] will point
to Docker Hub, and thus images are pushed there.

This fixes a bug where Octez Docker Distribution would be pushed to
the GCP registry on both protected and non-protected refs. What we
would like to have is:
 - publish images produced on the master pipeline to Docker Hub
 - publish images produced on non-protected refs to GCP registry.

And this commit should ensure this behavior.