EVM Node: Allow the proxy node to ignore block parameters
Without this option, it becomes pretty hard to use the proxy node with tools
like MetaMask which sets a block hash for most their requests. The time between
fetching the latest block and submitting requests is large enough that a new
Tezos block can be minted, which will make the requets fails.

Ignoring block parameter does not feel safe in the general case, but it can be
a useful tool for controled setup, where we understand well enough the behavior
of the caller.