Tezt/Cloud: Fix DNS registration when deploying with proxy mode
When deploying with proxy mode, the `add` transaction would fail if a
domain was already registered (but this won't fail the experiment).
By detecting first if an IP address already exists and by remove it we
should ensure this should not happen.

This should ensure that when running a new experiment using the proxy
mode, the domain should be registered correctly.