RISC-V: move address translation up the stack
Required as part of the work for the block cache. We want to be able to
fetch a block from a physical address - and if it fails to then do the
fetch/run_instr cycle.

To avoid re-doing the translation for the second case, we do this interrupt checking/address translating further up the stack - allowing more
simple delegation between blocks vs instructions.