EVM Node: Fix `observer.rollup_node_tracking` being ignored
Did I mention I don’t like Clic very much?

So, when testing in production™, I realized that `observer.rollup_node_tracking
= false` was happily ignored by the EVM node. The companion CLI flag
`--dont-track-rollup-node` is handled correctly, though.

The root of this issue lies in how Clic deals with CLI flag and how we deal
with configuration default. On the one hand, Clic assumes a flag is false
except if it is set. It does not advertize the absence of the flag with `None`.
On the other hand, we heavily rely on options to know when to fallback to the
default value. See where these two sentences do not mix very well together?

The proposed patch is not a silver bullet, but do indeed fixes the issue. We
use the opportunity to add a dedicated event to advertize the fact that the
rollup node tracking is disabled. This allows to write a proper test.