Lib_version: Versioning is for the EVM node
From a release perspective, our product is the `octez-evm-node` binary. It can
be used for Etherlink as for any other instance of our EVM-compatible Layer 2
blockchain.

In this patch, we propose to drop the versioning of Etherlink and focus on the
versioning of our binary instead. We might need a versioning scheme for the
kernel at some point, but it’s not the case at the moment, so let’s focus on
our immediate needs instead.

Here is the output of `octez-evm-node` printing its version, for reference.

    ; ./octez-evm-node --version
    1e89e2b2 (2024-08-21 18:08:03 +0200) (octez-evm-node 0.1)

A consequence of this patch is to deprecate Octez version tags not using the
`v` prefix. That is, `octez-22.0` is no longer a valid tag for an Octez
release, while it was before. This is to be able to easily distinguish between
a tag for Octez (`octez-v*`) and a tag for the Octez EVM node
(`octez-evm-node-v`). To be noted, tags like `octez-22.0` were not compatible
with the CI release pipeline as far as I can tell.