CI: Add `octez-evm-node` release pipeline
Hats off to everyone involved in the CIAO project. The result of your hard work
is true gem.

In a nutshell, we rewrite the Etherlink release logic to make use of the newly
introduced `octez-evm-node` product. Our goal is to introduce a specific
pipeline for releases of the Octez EVM Node (`octez-evm-node`), with two kinds
of artifacts:

  1. Docker images only containing the `octez-evm-node` binary
  2. `octez-evm-node` static binaries for x86-64 and Amd64

In order to implement that, I have ~~stolen~~ drawn inspiration from the Octez
release pipeline. Some scripts had to be generalized a little, because they
were assuming `octez-node` nad `octez-client` would always be part of Docker
images built by the CI. Some general-purpose functions were duplicated, like
the one to upload an artifact to a Gitlab package. This is to keep the
“official” release pipeline as simple as possible, and avoid indirections when
possible.

You can review the release page generated by Gitlab for a fake tage
[here][release-page]. Can you guess how many trials I had to do before getting
something working as I wanted?

[release-page]: https://gitlab.com/nomadic-labs/tezos/-/releases/octez-evm-node-v0.11