EVM Node: Have `Blueprints_follower` use `get_blueprint_number` once
Catching-up with an outdated EVM node using the `Blueprints_follower` module
calls a helpers (`get_blueprint_number`) as many time as blocks between the
local and remote head. This is not really necessary, since once we know the
initial head, we can easily infer what is the next expected one.