EVM Node: EVM services for RPC mode don't rely on blueprints follower
We remove the mutable field next_blueprint_number from the read-only handler to
the context, and we instead read it from the store. This change simplifies the
logic, because it means it’s means the read-only context does not need to
follow the stream of blueprints to be functioning as expected.