Etherlink/Sequencer+Kernel: don't sign the blueprints sent to DAL
This commit changes the format of blueprints sent to the DAL and
parsed by the kernel. The change consists in removing the signature
part.

In the sequencer, the signature is still computed at preparation time
and dropped before converting to RLP.

In the kernel, the parsing function for unsigned blueprints is used
instead of the one for signed ones.