EVM Node: Add a helper command to check a config file
Before deploying a new configuration for the EVM node (or before deploying a
new version of the EVM node itself) it can be handy to determine with certainty
whether or not said configuration file is correct. This feature has been
lacking, which is unfortunate considering how many breaking changes v0.1
brought to the configuration file.

In this patchset, we introduce a simple `check config` command, which is an
already pretty interesting first step towards full-fledged configuration
validation. For now, `check config` is limited to assert the configuration file
satisfies the JSON schema output by `describe config`.