EVM Node: Fetch the remote head when it looks like we have caught-up
Right now, the EVM node fetches the head of its upstream EVM node endpoint
once, then if this head is in the post, it crawls it until it catch-up one
blueprint at a time. This was done because the streamed RPC makes the EVM node
hangs if too many blueprints are sent at once.

This works well if there is not too many blueprints created between the time
the remote head is fetched and the catch-up is allegedly done, with “too many“
being loosely defined here. But in the case of a replay from genesis for
instance, we might need several _days_ to catch-up (fyi, it looks like we can
replay a little more than 500K blueprints per day as of now). Fortunately,
since our nodes are restarted periodically, we might get away with it (the node
is refetched every time the connection is lost), but still, we need to improve
the logic here.

In this changeset, I propose to fetch the remote head once we have alledgedly
catch-up, and to keep crawling if actually the upstream EVM node has made
progress.