EVM Node: Exporting snapshots is compatible with --dont-track-rollup
The `l1_l2_relationship` table is not always populated. To get the current
number from the store, it is best to use the `context_hashes` table. This way,
it becomes possible to export a snapshot from the data-dir of an EVM node which
has not run already, or that was always run with `--dont-track-rollup-node`