EVM Node: Expose the snapshot commands in the `man`
It is finally time to unveil to the world that our flagship feature for the
v0.2 of the EVM node is ready to be used! Well, not without taking this last
opportunity to refine a little the UX, by simplifying it. Instead of having two
redundant yet distinct command, we keep one with a snaphsot file path as an
optional argument.