EVM Node: Set journal mode of the snapshot database on export
By default, SQLite does not use WAL as its journal mode, but the Octez EVM node
heavily relies on it for making concurrent access to it. As a consequence, the
node will systematically try to set the journal mode to WAL if it is not the
case, even if the db was open in a read-only mode (it would be dangerous to run
a RPC node along a read-write node without WAL enabled anyway).

Trying to export a snapshot from a data-dir freshly created through an import
of a snapshot has shown that, by default, `VACUUM` does not propagate the
journal mode. As a consequence, the db created by `snapshot import` cannot be
used by a read-only process (typically to export another snapshot).

This MR modifies `Evm_store.vacuum` to fix this, by opening a connection to the
newly created db in order to set its journal mode as expected.