CI: Put all sanity jobs on critical path for dependent jobs
Jobs that execute out-of-stage-order (i.e. jobs that have [needs:])
will not wait on the jobs in the stage [sanity]. This commit ensures
that jobs that depend only on [trigger] also depend on all sanity
jobs. Since all dependent jobs should transitively depend on
[trigger], this should ensure that all dependent jobs transitively
depend on the sanity jobs as well.