Tezt cloud: expose Dal_node.create_from_endpoint in Agent module
The function Dal_node.create_from_endpoint can be used to create a
value of type Dal_node.t without having a Node.t at hand, the L1 node
to put in configuration is provided as an Endpoint.t.

This will be useful to create a fake Dal_node.t representing the
reverse proxy.