EVM Node: Fix the `octez_evm_node_head` metrics for the RPC node
Prior the this changeset, the metrics is only set in the `apply_blueprint`
routine, something the RPC mode never calls. We propose to add it to the main
event loop of the RPC mode as well. We don’t fix its companion
`octez_evm_node_confirmed_head` for now, as it’s less interesting for the RPC
mode (it does not influence the behavior of the RPC mode).