EVM Node: Remove it from the binaries distributed with Octez
Since we are distributing the Octez EVM Node using a dedicated release channel,
there is no reason to keep it in the official Octez suite. It is actually
likely to be more confusing than necessary.

We keep it as experimental so that the binary remains in the build artifacts
(Docker images, static binaries, etc.).

(cherry picked from commit 029caaa5410a9652d6f9c4801f84dff492cda6a8)