Etherlink+DAL: fix a bug preventing the use of slot index 0
The encode_u8 function which was used to encode slot indices, returned
a byte sequence of length 1 containing the byte \000. This is not
compatible with the RLP library used in the kernel which expects the
number 0 to be encoded by the empty byte sequence.

The Rlp.encode_int function behaves as expected so we simply use it instead.