RISC-V: run all jstz-bench messages in a single level
This is step one towards running the jstz-benchmark in the WASM debugger. In order to compare a like-for-like (where there is a GC slowdown over time), we want to execute all messages in a single level, to avoid the WASM debugger clearing down memory on every message (and thus avoiding the performance impact).

Since a kernel is forced to reboot on every level, we remove the need for this by running all messages in the same level.