EVM Node: Fix level-related metrics at startup
When monitoring an EVM node through its metrics, it is frequent to see the
latest and confirmed levels set to 0 on startup. This is because the metrics
were not systematically set ahead of starting the RPC server. This changeset
fixes this by (1) moving the initialization of metrics to `Evm_context` so that
the sequencer, observer and TE sequencer have that value correct, and (2)
fixing the logic directly for the RPC node, which does not start the
`Evm_context` worker.