Tezt cloud: handle commitment publications inside batches
To compute attestation ratios for bakers, we parse the operation
receipts related to DAL: DAL payloads in attestations and DAL
commitment publications. This commit fixes a bug affecting the parsing
of commitment publications when they are not the first operation of
their batch, which can happen when the commitments are published by
the rollup node in the same batch as a "smart_rollup_add_messages" (to
send a DAL import signal through the rollup inbox).