Rust: Default to 1.78 and update `rust_deps`
Alpine 3.20 is packaging Rust 1.78. Prior to having `rustup` available in the
CI, we were using the same Rust version as Alpine. Considering Rust 1.74 is
almost one year old (November 2023), it looks like a good idea to upgrade.

Besides, we run `cargo update` in `src/rust_deps` which notably means we start
using `ocaml-rs` 1.0.1 (instead of relying on the beta). This version seems to
support OCaml 5, which was not the case of the beta as far as I can tell.