EVM Node: Callbacks to implement host functions in C-compatible language
This change is proposed as part of the Octez EVM node’s WASM runtime
experimental feature. The goal of this project is to introduce an ad-hoc WASM
runtime, that is an alternative take of the Fast Execution runtime implemented
for the rollup node.

This commit adds the low-level functions the WASM runtime implemented in Rust
will have to use to implement the host functions (as well as some low-level
manipulations of the Irmin tree to implement cleanly the PVM life cycle).

In order to keep the code simpler and more straightforward, we don’t reuse the
libraries developed for the WASM PVM (in particular, `tree-encoding` and
`lazy-containers`). These libraries make sense in the context of the WASM PVM,
but would add an unnecessary overhead here.