Tezt: Improve the “sequencers diverge” scenario
This test was one of the most flaky in our test-suit, and the root cause was
hidden in the following log:

    [observer_evm_node5] Error:
    [observer_evm_node5]   Did not find service: GET http://127.0.0.1:45379/evm/blueprint/0
    observer_evm_node5 exited with code 1.

This was preventing the following snippet to be resolved

    and* _ = Evm_node.wait_for_diverged observer
    and* _ = Evm_node.wait_for_shutdown_event ~can_terminate:true observer

Eventually triggering a timeout.

The second sequencer was initialized with `init from rollup node`, and the
assumption was that the rollup node wouldn’t have seen any block before that
time. As it turns out, the assumption was false.

This patch proposes another approach to initialize the second sequnecer:
exporting a snapshot from the first one and importing it. This should fix the
flakiness.