EVM Node: Generalize the `--finalized-view` CLI argument
In a previous episode, we’ve added `--finalized-view` to the proxy mode to
provide a lightweight node whose main purpose would be to only expose to its
users the “finalized” block of an Etherlink instance.

With this changeset, we generalize this mode to make it work with all the EVM
node modes. In such a mode and as far as RPC are concerned, the `latest` block
parameter becomes a synonym for `finalized`. We deprecate the previous option,
add a warning to make it clear it will be removed in a future release, and we
update the existing test.