RISC-V/jstz: reset gc between kernel runs
We were seeing a slowdown over time of the speed at which jstz was handling transactions.
This is due to boa gc (for some reason) being unable to garbage collect after each run.

For now, we manually drop the garbage collector between runs to force an entire reset. This results in TPS
becoming much more stable